Moved the new router to a separate file
