+++ src/components/app/details/appConfig/appConfig.scss
@@ -203,7 +203,7 @@
     }
 }
 
-.new-navigation-router {
+.cm-secret-list-router {
     &__header {
         &__back-cta {
             background: var(--N50);
